+++ examples/nextjs13/app/layout.tsx
@@ -1,3 +1,10 @@
+import { Inter } from "next/font/google";
+import "@upstash/react-cli/dist/index.css";
+
+const inter = Inter({
+  subsets: ["latin"],
+});
+
 export default function RootLayout({
   children,
 }: {
@@ -10,7 +17,7 @@ export default function RootLayout({
         head.tsx. Find out more at https://beta.nextjs.org/docs/api-reference/file-conventions/head
       */}
       <head />
-      <body>{children}</body>
+      <body className={inter.className}>{children}</body>
     </html>
   );
 }