+++ examples/with-tailwind/tailwind.config.js
@@ -1,8 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ["./app/**/*.tsx", "./node_modules/@upstash/react-cli/dist/*.js"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};