+++ examples/nextjs13/CHANGELOG.md
@@ -1,5 +1,12 @@
 # nextjs13
 
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @upstash/react-cli@1.0.0
+
 ## 0.1.10
 
 ### Patch Changes